% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_administrador.R
\name{dt_adm_gen_fechas}
\alias{dt_adm_gen_fechas}
\title{Descargar los datos adm_gen_fechas}
\usage{
dt_adm_gen_fechas(conexion, periodo_analisis)
}
\arguments{
\item{conexion}{clase formal. Conexi칩n base de datos}

\item{periodo_analisis}{clase array date. Debe contener la fecha inicio y fin del an치lisis}
}
\description{
Esta funci칩n descarga los datos de la tabla adm_gen_fechas para un periodo de an치lisis
}
