% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_administrador.R
\name{dt_adm_gen_miembros}
\alias{dt_adm_gen_miembros}
\title{Descargar los datos adm_gen_miembros}
\usage{
dt_adm_gen_miembros(conexion, segmentos_analisis = NULL, seudonimo = "REAL")
}
\arguments{
\item{conexion}{clase formal. Conexi贸n base de datos}

\item{segmentos_analisis}{clase array character. Lista de segmentos ("GE","CD","CV","C2","C7","C8","C9")
de los cuales se desea descargar la informaci贸n. Por defecto descarga la informaci贸n de todos los segmentos.}

\item{seudonimo}{clase character. Debe ser igual a "REAL" o "FICTICIO".Por defecto "REAL"}
}
\description{
Esta funci贸n descarga los datos de la tabla adm_gen_miembros
con base en los parametros ingresados
}
