% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_administrador.R
\name{dt_adm_gen_perfiles}
\alias{dt_adm_gen_perfiles}
\title{Descargar los datos adm_gen_perfiles}
\usage{
dt_adm_gen_perfiles(conexion, perfil_analisis)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{perfil_analisis}{clase character. Nombre del perfil}
}
\description{
Esta función descarga los datos de la tabla adm_gen_perfiles para un perfil
}
