% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_administrador.R
\name{dt_adm_gen_segmentos}
\alias{dt_adm_gen_segmentos}
\title{Descarga los datos adm_gen_segmentos}
\usage{
dt_adm_gen_segmentos(conexion)
}
\description{
Esta función descarga los datos de la tabla adm_gen_segmentos
}
