% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_app.R
\name{dt_app_gen_estructura_llave}
\alias{dt_app_gen_estructura_llave}
\title{Descargar los datos app_gen_estructura_llave}
\usage{
dt_app_gen_estructura_llave(conexion, llave_analisis, segmento_analisis)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{llave_analisis}{clase character.}

\item{segmento_analisis}{clase character. Segmento analisis ("GE","CD","CV","C2","C7","C8","C9")}
}
\description{
Esta función descarga los datos de la tabla app_gen_estructura_llave
con base en los parametros ingresados
}
