% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_general.R
\name{dt_cuenta_garantia_identificacion_seudonimo_sql}
\alias{dt_cuenta_garantia_identificacion_seudonimo_sql}
\title{Devuelve el sql para extraer CUENTA_GARANTIA_IDENTIFICACION_SEUDONIMO}
\usage{
dt_cuenta_garantia_identificacion_seudonimo_sql(seudonimo)
}
\arguments{
\item{seudonimo}{clase character. Debe ser igual a "REAL" o "FICTICIO".}
}
\description{
Esta función devuelve el sql para extraer la CUENTA_GARANTIA_IDENTIFICACION_SEUDONIMO acorde a la condición.
}
