% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_derivados.R
\name{dt_dv_curva_fwd}
\alias{dt_dv_curva_fwd}
\title{Descarga los datos dv_curva_fwd}
\usage{
dt_dv_curva_fwd(
  conexion,
  proveedor = "MySQL",
  periodo_analisis = NULL,
  fecha_analisis = NULL
)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{proveedor}{clase character. Proveedor de la base de datos ("Oracle", "MySQL"). Por defecto "MySQL"}

\item{periodo_analisis}{clase array date. Debe contener la fecha inicio y fin del análisis}

\item{fecha_analisis}{clase date. Debe contener la fecha del análisis, si el parametro periodo_analisis es
diferente de NULL este parametro no se tendra en cuenta. Por defecto NULL}
}
\description{
Esta función descarga los datos de la tabla dv_curva_fwd para un periodo de análisis y
con base en los parametros ingresados
}
