% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_derivados.R
\name{dt_dv_pa_por_rango}
\alias{dt_dv_pa_por_rango}
\title{Descarga los datos dv_pa_por_rango}
\usage{
dt_dv_pa_por_rango(
  conexion,
  periodo_analisis = NULL,
  fecha_analisis = NULL,
  ficticio = FALSE
)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{periodo_analisis}{clase array date. Debe contener la fecha inicio y fin del análisis}

\item{fecha_analisis}{clase date. Debe contener la fecha del análisis, si el parametro periodo_analisis es
diferente de NULL este parametro no se tendra en cuenta. Por defecto NULL}

\item{ficticio}{clase boolean. TRUE si se desea que el "ID_SEUDONIMO" de los miembros se igua al "ID_FICTICIO"  en
caso contrario sera igual al "ID". Por defecto FALSE}
}
\description{
Esta función descarga los datos de la tabla dv_pa_por_rango para un periodo de análisis y
con base en los parametros ingresados
}
