% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_general.R
\name{dt_fecha_analisis_sql}
\alias{dt_fecha_analisis_sql}
\title{Convertir fecha an치lisis a formato SQL (system)}
\usage{
dt_fecha_analisis_sql(fecha_analisis, oracle = FALSE)
}
\arguments{
\item{fecha_analisis}{clase date. Debe contener la fecha del an치lisis}

\item{oracle}{clase boolean. TRUE si la base de datos utilizada es oracle. Por defecto FALSE}
}
\description{
Esta funci칩n convierte una fecha de an치lisis a formato SQL dependiendo del tipo de base de datos
}
