% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_general.R
\name{dt_ficticio_sql}
\alias{dt_ficticio_sql}
\title{ID Miembros SQL (system)}
\usage{
dt_ficticio_sql(ficticio)
}
\arguments{
\item{ficticio}{clase boolean. TRUE si se desea descargar el 'ID_FICTICIO' de los miembros en
caso contrario se descargara el 'ID'. Por defecto FALSE}
}
\description{
Esta función devuelve el ID correspondiente acorde a la condición.
}
