% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_general.R
\name{dt_filtro_datos}
\alias{dt_filtro_datos}
\title{Filtrar el data.frame con base en los inputs}
\usage{
dt_filtro_datos(
  datos,
  fecha_analisis = NULL,
  segmentos = NULL,
  miembros = NULL,
  cuentas = NULL,
  activos = NULL
)
}
\arguments{
\item{datos}{clase data.frame. Los datos a filtrar}

\item{fecha_analisis}{clase date. Debe contener la fecha del análisis}

\item{segmentos}{clase array character. Debe contener los segmentos que se desean filtrar. Por defecto NULL}

\item{miembros}{clase array character. Debe contener los miembros que se desean filtrar. Por defecto NULL}

\item{cuentas}{clase array character. Debe contener los cuentas que se desean filtrar. Por defecto NULL}
}
\description{
Esta función filtra el data.frame con base en los inputs segmentos, miembros y cuentas
}
