% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_collateral_management.R
\name{dt_gen_cm_acciones}
\alias{dt_gen_cm_acciones}
\title{Descarga los datos gen_cm_acciones}
\usage{
dt_gen_cm_acciones(
  conexion,
  periodo_analisis = NULL,
  fecha_analisis = NULL,
  seudonimo = "REAL"
)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{periodo_analisis}{clase array date. Debe contener la fecha inicio y fin del análisis}

\item{fecha_analisis}{clase date. Debe contener la fecha del análisis, si el parametro periodo_analisis es
diferente de NULL este parametro no se tendra en cuenta. Por defecto NULL}

\item{seudonimo}{clase character. Debe ser igual a "REAL" o "FICTICIO".Por defecto "REAL"}
}
\description{
Esta función descarga los datos de la tabla gen_cm_acciones para un periodo de análisis y
con base en los parametros ingresados. Adicionalmente tranforma la tabla (Cambia su estructura a pivot_longer)
}
