% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_collateral_management.R
\name{dt_gen_cm_promedio_diario}
\alias{dt_gen_cm_promedio_diario}
\title{Promedio diario de los datos gen_cm_titulos o gen_cm_acciones}
\usage{
dt_gen_cm_promedio_diario(datos)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_cm_titulos}} / \code{\link{dt_gen_cm_acciones}} o tener una estructura igual a dichos datos}
}
\description{
Esta función calcula el promedio diario de los datos gen_cm_titulos o gen_cm_acciones.
}
