% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_collateral_management.R
\name{dt_gen_gar_dep_resumen}
\alias{dt_gen_gar_dep_resumen}
\title{Descarga los datos gen_gar_depo_resumen}
\usage{
dt_gen_gar_dep_resumen(
  conexion,
  periodo_analisis = NULL,
  fecha_analisis = NULL,
  miembros_analisis = NULL,
  segmentos_analisis = NULL,
  seudonimo = "REAL"
)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{periodo_analisis}{clase array date. Debe contener la fecha inicio y fin del análisis}

\item{fecha_analisis}{clase date. Debe contener la fecha del análisis, si el parametro periodo_analisis es
diferente de NULL este parametro no se tendra en cuenta. Por defecto NULL}

\item{miembros_analisis}{clase array character. Lista de miembros
de los cuales se desea descargar la información. Por defecto descarga la información de todos los miembros}

\item{segmentos_analisis}{clase array character. Lista de segmentos ("GE","CN","CV","C2","C7","C8","C9")
de los cuales se desea descargar la información. Por defecto descarga la información de todos los segmentos.}

\item{seudonimo}{clase character. Debe ser igual a "REAL" o "FICTICIO".Por defecto "REAL"}
}
\description{
Esta función descarga los datos de la tabla gen_gar_depo_resumen para un periodo de análisis y
con base en los parametros ingresados
}
