% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ingresos.R
\name{dt_gen_ing_cumplimiento_presupuesto}
\alias{dt_gen_ing_cumplimiento_presupuesto}
\title{Crea la tabla gen_ing_cumplimiento_presupuesto}
\usage{
dt_gen_ing_cumplimiento_presupuesto(
  conexion,
  datos,
  periodo_analisis = NULL,
  fecha_analisis = NULL
)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_ing_resumen}} o tener una estructura igual a dichos datos}

\item{periodo_analisis}{clase array date. Debe contener la fecha inicio y fin del análisis}

\item{fecha_analisis}{clase date. Debe contener la fecha del análisis, si el parametro periodo_analisis es
diferente de NULL este parametro no se tendra en cuenta. Por defecto NULL}
}
\description{
Esta función descarga los datos params_proyeccion_ingresos y los une a gen_ingresos_resumen (descargados con anterioridad)
creando la tabla gen_ing_cumplimiento_presupuesto (localmente).
}
