% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_garantia.R
\name{dt_gen_lmc_consumo}
\alias{dt_gen_lmc_consumo}
\title{Descarga los datos gen_lmc_consumo}
\usage{
dt_gen_lmc_consumo(
  conexion,
  periodo_analisis = NULL,
  fecha_analisis = NULL,
  miembros_liq_analisis = NULL,
  seudonimo = "REAL"
)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{periodo_analisis}{clase array date. Debe contener la fecha inicio y fin del análisis. Por defecto NULL}

\item{fecha_analisis}{clase date. Debe contener la fecha del análisis, si el parametro periodo_analisis es
diferente de NULL este parametro no se tendra en cuenta. Por defecto NULL}

\item{miembros_liq_analisis}{clase array character. Lista de miembros liquidadores
de los cuales se desea descargar la información. Por defecto descarga la información de todos los miembros liquidadores}

\item{seudonimo}{clase character. Debe ser igual a "REAL" o "FICTICIO".Por defecto "REAL"}
}
\description{
Esta función descarga los datos de la tabla gen_lmc_consumo para un periodo de análisis y
con base en los parametros ingresados
}
