% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_posicion_abierta.R
\name{dt_gen_pa_resumen}
\alias{dt_gen_pa_resumen}
\title{Descarga los datos gen_pa_resumen}
\usage{
dt_gen_pa_resumen(
  conexion,
  proveedor = "MySQL",
  periodo_analisis = NULL,
  fecha_analisis = NULL,
  segmentos_analisis = NULL,
  ficticio = FALSE,
  valorada = TRUE,
  importe = FALSE
)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{proveedor}{clase character. Proveedor de la base de datos ("Oracle", "MySQL"). Por defecto "MySQL"}

\item{periodo_analisis}{clase array date. Debe contener la fecha inicio y fin del análisis}

\item{fecha_analisis}{clase date. Debe contener la fecha del análisis, si el parametro periodo_analisis es
diferente de NULL este parametro no se tendra en cuenta. Por defecto NULL}

\item{segmentos_analisis}{clase array character. Lista de segmentos ("GE","CN","CV","C2","C7","C8","C9")
de los cuales se desea descargar la información. Por defecto descarga la información de todos los segmentos.}

\item{ficticio}{clase boolean. TRUE si se desea que el "ID_SEUDONIMO" de los miembros se igua al "ID_FICTICIO"  en
caso contrario sera igual al "ID". Por defecto FALSE}

\item{valorada}{clase boolean. TRUE si se desea descargar la posición abierta valorada. Por defecto TRUE}

\item{importe}{clase boolean. TRUE si se desea descargar la posición abierta importe. Por defecto FALSE}
}
\description{
Esta función descarga los datos de la tabla gen_pa_resumen para un periodo de análisis y
con base en los parametros ingresados
}
