% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_general.R
\name{dt_id_seudonimo}
\alias{dt_id_seudonimo}
\title{ID_SEUDONIMO Miembros}
\usage{
dt_id_seudonimo(seudonimo)
}
\arguments{
\item{seudonimo}{clase character. Debe ser igual a "REAL" o "FICTICIO".}
}
\description{
Esta función devuelve el ID_SEUDONIMO correspondiente acorde a la condición.
}
