% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_general.R
\name{dt_nombre_abreviacion_seudonimo}
\alias{dt_nombre_abreviacion_seudonimo}
\title{NOMBRE_SEUDONIMO Miembros}
\usage{
dt_nombre_abreviacion_seudonimo(seudonimo)
}
\arguments{
\item{seudonimo}{clase character. Debe ser igual a "REAL" o "FICTICIO".}
}
\description{
Esta función devuelve el NOMBRE_ABREVIACION_SEUDONIMO correspondiente acorde a la condición.
}
