% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_general.R
\name{dt_num_char}
\alias{dt_num_char}
\title{Convertir numero a caracter (system)}
\usage{
dt_num_char(x)
}
\arguments{
\item{x}{clase column. Columna sobre la cual se va a realizar la converción}
}
\description{
Esta función convierte un numero a caracter. Unicamente se puede aplicar a un data.frame
}
