% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_parametros.R
\name{dt_pa_div_lineas_pl}
\alias{dt_pa_div_lineas_pl}
\title{Descarga los datos pa_div_lineas_pl}
\usage{
dt_pa_div_lineas_pl(conexion, seudonimo = "REAL")
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{seudonimo}{clase character. Debe ser igual a "REAL" o "FICTICIO".Por defecto "REAL"}
}
\description{
Esta función descarga los datos de la tabla pa_div_lineas_pl
}
