% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_general.R
\name{dt_periodo_analisis_sql}
\alias{dt_periodo_analisis_sql}
\title{Convertir periodo an치lisis a formato SQL (system)}
\usage{
dt_periodo_analisis_sql(periodo_analisis)
}
\arguments{
\item{periodo_analisis}{clase array date. Debe contener la fecha inicio y fin del an치lisis}
}
\description{
Esta funci칩n convierte un periodo de an치lisis a formato SQL dependiendo del tipo de base de datos
}
