% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_general.R
\name{dt_porcentaje_caracter}
\alias{dt_porcentaje_caracter}
\title{Convertir un numero a porcentaje (system)}
\usage{
dt_porcentaje_caracter(x, digitos = 1)
}
\arguments{
\item{x}{clase numeric. Valor a convertir}

\item{digitos}{clase integer. Número de digitos a contemplar en el porcentaje}
}
\description{
Esta función convierte un numero a porcentaje retornando el valor como un caracter
}
