% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_general.R
\name{dt_porcentaje_variacion}
\alias{dt_porcentaje_variacion}
\title{Calcular porcentaje de variaci贸n (system)}
\usage{
dt_porcentaje_variacion(x)
}
\arguments{
\item{x}{clase column. Columna sobre la cual se va a calcular la variaci贸n}
}
\description{
Esta funci贸n calcula el porcentaje de variaci贸n de una variable entre dos filas
consecutivas. Unicamente se puede aplicar a un data.frame
}
