% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_renta_fija.R
\name{dt_rf_curva_tes}
\alias{dt_rf_curva_tes}
\title{Descarga los datos rf_curva_tes}
\usage{
dt_rf_curva_tes(
  conexion,
  proveedor = "MySQL",
  periodo_analisis = NULL,
  fecha_analisis = NULL
)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{periodo_analisis}{clase array date. Debe contener la fecha inicio y fin del análisis}

\item{fecha_analisis}{clase date. Debe contener la fecha del análisis, si el parametro periodo_analisis es
diferente de NULL este parametro no se tendra en cuenta. Por defecto NULL}
}
\description{
Esta función descarga los datos de la tabla rf_curva_tes para un periodo de análisis y
con base en los parametros ingresados
}
