% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_renta_variable.R
\name{dt_rv_haircuts}
\alias{dt_rv_haircuts}
\title{Descarga los datos rv_haircuts}
\usage{
dt_rv_haircuts(conexion, periodo_analisis = NULL, fecha_analisis = NULL)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{periodo_analisis}{clase array date. Debe contener la fecha inicio y fin del análisis}

\item{fecha_analisis}{clase date. Debe contener la fecha del análisis, si el parametro periodo_analisis es
diferente de NULL este parametro no se tendra en cuenta. Por defecto NULL}
}
\description{
Esta función descarga los datos de la tabla rv_haircuts para un periodo de análisis y
con base en los parametros ingresados
}
