% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_renta_variable.R
\name{dt_rv_pa_repos}
\alias{dt_rv_pa_repos}
\title{Descarga los datos gen_pa para el producto repos}
\usage{
dt_rv_pa_repos(
  conexion,
  periodo_analisis = NULL,
  fecha_analisis = NULL,
  seudonimo = "REAL",
  nominal = FALSE,
  valorada = TRUE,
  importe = FALSE
)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{periodo_analisis}{clase array date. Debe contener la fecha inicio y fin del análisis}

\item{fecha_analisis}{clase date. Debe contener la fecha del análisis, si el parametro periodo_analisis es
diferente de NULL este parametro no se tendra en cuenta. Por defecto NULL}

\item{seudonimo}{clase character. Debe ser igual a "REAL" o "FICTICIO".Por defecto "REAL"}

\item{nominal}{clase boolean. TRUE si se desea descargar la posición abierta nominal Por defecto TRUE}

\item{valorada}{clase boolean. TRUE si se desea descargar la posición abierta valorada. Por defecto TRUE}

\item{importe}{clase boolean. TRUE si se desea descargar la posición abierta importe. Por defecto FALSE}
}
\description{
Esta función descarga los datos de la tabla gen_pa para el producto repos, un periodo de análisis y
con base en los parametros ingresados
}
