% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_general.R
\name{dt_segmentos_analisis_sql}
\alias{dt_segmentos_analisis_sql}
\title{Convertir segmentos_analisis a formato SQL (system)}
\usage{
dt_segmentos_analisis_sql(segmentos_analisis, proveedor = "MySQL")
}
\arguments{
\item{segmentos_analisis}{clase array character. Debe contener la lista de segmentos análisis}

\item{proveedor}{clase character. Proveedor de la base de datos ("Oracle", "MySQL"). Por defecto "MySQL"}
}
\description{
Esta función convierte una lista de segmentos a formato SQL
}
