% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_divisas.R
\name{gt_div_dist_max_obligacion_diaria}
\alias{gt_div_dist_max_obligacion_diaria}
\title{Grafica la distribución de la maxima obligación diaria por miembro (boxplot)}
\usage{
gt_div_dist_max_obligacion_diaria(datos, colores, fixedrange = FALSE)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_div_analisis_lole}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función \code{\link{dt_adm_gen_colores}}}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}
}
\description{
Esta función crea la gráfica de la distribución de la maxima obligación diaria por miembro en formato de boxplot
}
