% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_monitoreo_comportamiento.R
\name{gt_div_monitoreo_comportamiento_diario_miembro}
\alias{gt_div_monitoreo_comportamiento_diario_miembro}
\title{Grafica los datos atípicos de un miembro sobre el volumen, liquidación, número operaciones diaras y participación (lineas+puntos)}
\usage{
gt_div_monitoreo_comportamiento_diario_miembro(datos, fixedrange = FALSE)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_div_monitoreo_comportamiento}} o tener una estructura igual a dichos datos
(Solo deben incluir los datos de un miembro)}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}
}
\description{
Esta función crea la gráfica de los datos atípicos de un miembro sobre el volumen (eje y1),
liquidación (eje y2), número operaciones diarias (eje y3) y participación (eje y4) en formato de lineas y puntos.
}
