% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_monitoreo_comportamiento.R
\name{gt_div_monitoreo_comportamiento_diario_por_miembro}
\alias{gt_div_monitoreo_comportamiento_diario_por_miembro}
\title{Grafica los datos atípicos diarios por miembro (puntos)}
\usage{
gt_div_monitoreo_comportamiento_diario_por_miembro(datos, fixedrange = FALSE)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_div_monitoreo_comportamiento}} o tener una estructura igual a dichos datos
(Solo deben incluir los datos de un miembro)}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}
}
\description{
Esta función crea la gráfica de loslos datos atípicos diarios por miembro en formato de puntos.
}
