% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_monitoreo_comportamiento.R
\name{gt_div_monitoreo_comportamiento_miembro}
\alias{gt_div_monitoreo_comportamiento_miembro}
\title{Grafica el volumen, liquidación, número operaciones diaras y participación
de un miembro para una fecha de análisis vs. su comportamiento típico (radar)}
\usage{
gt_div_monitoreo_comportamiento_miembro(datos, fecha_analisis)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_div_monitoreo_comportamiento}} o tener una estructura igual a dichos datos
(Solo deben incluir los datos de un miembro)}

\item{fecha_analisis}{clase date. Fecha en la que se realiza el análisis (Último día de los datos)}
}
\description{
Esta función crea la gráfica del volumen, liquidación, número operaciones diaras y participación
de un miembro para una fecha de análisis vs. su comportamiento típico (radar) en formato de radar.
}
