% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_monitoreo_comportamiento.R
\name{gt_div_monitoreo_comportamiento_resumen}
\alias{gt_div_monitoreo_comportamiento_resumen}
\title{Tabla monitoreo comportamiento resumen}
\usage{
gt_div_monitoreo_comportamiento_resumen(
  datos,
  fecha_analisis,
  pageLength = 100,
  style = "bootstrap4"
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_div_monitoreo_comportamiento}} o tener una estructura igual a dichos datos}

\item{fecha_analisis}{clase date. Fecha en la que se realiza el análisis (Último día de los datos)}

\item{pageLength}{clase number. Número de filas por hoja que alojara
la tabla. Por defecto 100}

\item{style}{clase character. Estilo boostrap que se debe utilizar
para renderizar la tabla. Por defecto "bootstrap4"}
}
\description{
Esta función crea la tabla monitoreo comportamiento en formato html
}
