% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_divisas.R
\name{gt_div_vol_liq_op_por_miembro}
\alias{gt_div_vol_liq_op_por_miembro}
\title{Grafica el volumen vs. liquidación vs. número operaciones por miembro (barras)}
\usage{
gt_div_vol_liq_op_por_miembro(
  datos,
  colores,
  fixedrange = FALSE,
  boton_activo = NULL,
  botones_inactivos = c()
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_div_vol_liq_op}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función \code{\link{dt_adm_gen_colores}}}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico. Por defecto NULL}

\item{botones_inactivos}{clase vector character. Vector de los nombres de los botones a desactivar
en la gráfica ("Segmento"). Por defecto c()}
}
\description{
Esta función crea la gráfica del volumen (eje y1) vs. liquidación (eje y2) vs. número operaciones por miembro
en formato de barras.
La información se muestra acorde a la agrupación relacionada con cada botón
}
