% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_divisas.R
\name{gt_div_vol_liq_op_promedio_diario}
\alias{gt_div_vol_liq_op_promedio_diario}
\title{Grafica el volumen vs. liquidación vs. número operaciones promedio diario por (Mes o Año) (barras)}
\usage{
gt_div_vol_liq_op_promedio_diario(
  datos,
  colores,
  fixedrange = FALSE,
  dash_board = TRUE,
  promedio = "m",
  boton_activo = NULL,
  botones_inactivos = c()
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_pa_resumen}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función \code{\link{dt_adm_gen_colores}}}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{dash_board}{clase boolean. TRUE si se desea monstrar en la agrupaciones simetricas unicamente el eje positivo. Por defecto TRUE}

\item{promedio}{clase character. "m" si se desea promediar por mes y "y" si se desea promediar por año. Por defecto "m"}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico. Por defecto NULL}

\item{botones_inactivos}{clase vector character. Vector de los nombres de los botones a desactivar
en la gráfica ("Segmento", "Tipo Miembro"). Por defecto c()}
}
\description{
Esta función crea la gráfica del volumen (eje y1) vs. liquidación (eje y2) vs. número operaciones promedio diario en formato de barras
La información se muestra acorde a la agrupación relacionada con cada botón
}
