% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_derivados.R
\name{gt_dv_indicador_liquidez}
\alias{gt_dv_indicador_liquidez}
\title{Grafica el indicador de liquidez  (barras)}
\usage{
gt_dv_indicador_liquidez(datos, fixedrange = FALSE)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_dv_indicador_liquidez}} o tener una estructura igual a dichos datos}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}
}
\description{
Esta función crea la gráfica del indicador de liquidez en formato de barras.
}
