% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_derivados.R
\name{gt_dv_liq_diarias}
\alias{gt_dv_liq_diarias}
\title{Grafica las liquidaciones de derivados diarias  (lineas)}
\usage{
gt_dv_liq_diarias(datos, fixedrange = FALSE)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_dv_liq_resumen}} o tener una estructura igual a dichos datos}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}
}
\description{
Esta función crea la gráfica de las liquidaciones de derivados diarias en formato de lineas.
}
