% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_derivados.R
\name{gt_dv_liq_por_miembro_liq}
\alias{gt_dv_liq_por_miembro_liq}
\title{Grafica las liquidaciones de derivados por miembro liquidador  (barras)}
\usage{
gt_dv_liq_por_miembro_liq(
  datos,
  colores,
  fixedrange = FALSE,
  boton_activo = NULL,
  botones_inactivos = c()
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_dv_liq_resumen}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función \code{\link{dt_adm_gen_colores}}}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico. Por defecto NULL}

\item{botones_inactivos}{clase vector character. Vector de los nombres de los botones a desactivar
en la gráfica ("Tipo Producto", "Subtipo Producto", "Origen Producto", "Tipo Cuenta Gar."). Por defecto c()}
}
\description{
Esta función crea la gráfica de las liquidaciones de derivados por miembro liquidador en formato de barras.
La información se muestra acorde a la agrupación relacionada con cada botón
}
