% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_derivados.R
\name{gt_dv_pa_neta_diaria_por_miembro}
\alias{gt_dv_pa_neta_diaria_por_miembro}
\title{Grafica la posición abierta neta diaria por miembro (heatmap)}
\usage{
gt_dv_pa_neta_diaria_por_miembro(
  datos,
  fixedrange = FALSE,
  boton_activo = NULL,
  botones_inactivos = c()
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_dv_pa_por_rango}} o tener una estructura igual a dichos datos}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico ("F. TES", "OIS", "TRM"). Por defecto NULL}
}
\description{
Esta función crea la gráfica de la posición abierta neta por miembro en formato de heatmap.
La información se muestra acorde a la agrupación relacionada con cada botón
}
