% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_derivados.R
\name{gt_dv_pa_neta_por_rango_miembro}
\alias{gt_dv_pa_neta_por_rango_miembro}
\title{Grafica la posición abierta neta por rango y miembro (barras)}
\usage{
gt_dv_pa_neta_por_rango_miembro(
  datos,
  colores,
  fixedrange = FALSE,
  boton_activo = NULL
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_dv_pa_por_rango}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función \code{\link{dt_colores}}}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico ("F. TES", "OIS", "TRM"). Por defecto NULL}
}
\description{
Esta función crea la gráfica de la posición abierta neta por rango y miembro en formato de barras.
La información se muestra acorde a la agrupación relacionada con cada botón
}
