% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_derivados.R
\name{gt_dv_pa_por_rango_resumen}
\alias{gt_dv_pa_por_rango_resumen}
\title{Tabla posición abierta por rango resumen}
\usage{
gt_dv_pa_por_rango_resumen(
  datos,
  fecha_analisis,
  pageLength = 100,
  style = "bootstrap4"
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_dv_pa_por_rango}} o tener una estructura igual a dichos datos}

\item{fecha_analisis}{clase date. Fecha en la que se realiza el análisis (Último día de los datos)}

\item{pageLength}{clase number. Número de filas por hoja que alojara
la tabla. Por defecto 100}

\item{style}{clase character. Estilo boostrap que se debe utilizar
para renderizar la tabla. Por defecto "bootstrap4"}
}
\description{
Esta función crea la posición abierta por rango en formato html
}
