% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_derivados.R
\name{gt_dv_reverse_gap_por_miembro}
\alias{gt_dv_reverse_gap_por_miembro}
\title{Grafica el reverse gap por miembro (barras)}
\usage{
gt_dv_reverse_gap_por_miembro(datos, fixedrange = FALSE, boton_activo = NULL)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_dv_pa_por_rango}} o tener una estructura igual a dichos datos}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico ("Vencimiento hasta 1 día", "Vencimiento hasta 2 días"). Por defecto NULL}
}
\description{
Esta función crea la gráfica del reverse gap por miembro en formato de barras
La información se muestra acorde a la agrupación relacionada con cada botón
}
