% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_general.R
\docType{data}
\name{gt_espanol}
\alias{gt_espanol}
\title{Retorna la lista de configuración del idioma de las data.table (DT)}
\format{
An object of class \code{list} of length 14.
}
\usage{
gt_espanol
}
\description{
Retorna la lista de configuración del idioma de las data.table (DT)
}
\keyword{datasets}
