% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_garantia.R
\name{gt_gar_dep_exi}
\alias{gt_gar_dep_exi}
\title{Gráfica la garantía depositada vs. exigida (barras h)}
\usage{
gt_gar_dep_exi(
  datos,
  colores,
  fixedrange = FALSE,
  boton_activo = NULL,
  botones_inactivos = c()
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_gar_dep_exi_resumen}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función \code{\link{dt_colores}}}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico. Por defecto NULL}

\item{botones_inactivos}{clase vector character. Vector de los nombres de los botones a desactivar
en la gráfica ("Segmento", "Tipo Miembro", "Tipo Cuenta Gar."). Por defecto c()}
}
\description{
Esta función crea la gráfica de la garantía depositada vs. exigida
en formato de barras horizantales.
La información se muestra acorde a la agrupación relacionada con cada botón
}
