% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_garantia.R
\name{gt_gar_dep_exi_promedio_diario}
\alias{gt_gar_dep_exi_promedio_diario}
\title{Gráfica la garantía depositada vs. exigida promedio diario (barras)}
\usage{
gt_gar_dep_exi_promedio_diario(
  datos,
  colores,
  fixedrange = FALSE,
  promedio = "m",
  boton_activo = NULL,
  botones_inactivos = c()
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_gar_dep_exi}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función \code{\link{dt_adm_gen_colores}}}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{promedio}{clase character. "m" si se desea promediar por mes y "y" si se desea promediar por año. Por defecto "m"}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico. Por defecto NULL}

\item{botones_inactivos}{clase vector character. Vector de los nombres de los botones a desactivar
en la gráfica. Si 'dos_ejes' es igual a true los botones son ("Segmento", "Tipo Miembro", "Tipo Cuenta Gar."). En caso
contrario los botones son (General,D-sss--s-s-s-s-s-s--s). Por defecto c()}
}
\description{
Esta función crea la gráfica de la garantía depositada (eje y1) vs. exigida (eje y2) promedio diario en formato de barras
La información se muestra acorde a la agrupación relacionada con cada botón
}
