% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_collateral_management.R
\name{gt_gar_dep_promedio_diario}
\alias{gt_gar_dep_promedio_diario}
\title{Gráfica la garantía depositada promedio diario por (Mes o Año) (barras)}
\usage{
gt_gar_dep_promedio_diario(datos, colores, fixedrange = FALSE, promedio = "m")
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_gar_dep_resumen_periodo}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
#' por la función colores}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{promedio}{clase character. "m" si se desea promediar por mes y "y" si se desea promediar por año. Por defecto "m"}
}
\description{
Esta función crea la gráfica de la garantía depositada promedio diario en formato de barras.
}
