% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_collateral_management.R
\name{gt_gar_dep_promedio_diario_por_miembro_accion_tipocuenta}
\alias{gt_gar_dep_promedio_diario_por_miembro_accion_tipocuenta}
\title{Gráfica la garantía depositada por miembro, acción y tipo cuenta promedio diario (treemap)}
\usage{
gt_gar_dep_promedio_diario_por_miembro_accion_tipocuenta(datos)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_cm_promedio_diario}} o tener una estructura igual a dichos datos}
}
\description{
Esta función crea la gráfica de la  garantía depositada promedio diario por miembro, acción y tipo cuenta en
formato treemap
}
