% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_collateral_management.R
\name{gt_gar_dep_ratio_liquidacion_por_titulo}
\alias{gt_gar_dep_ratio_liquidacion_por_titulo}
\title{Gráfica el ratio periodo de liquidación por título (heatmap)}
\usage{
gt_gar_dep_ratio_liquidacion_por_titulo(
  datos,
  fixedrange = FALSE,
  boton_activo = NULL
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_cm_titulos_periodo}} o tener una estructura igual a dichos datos}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico ("General", "MEC+SEN", "Simultaneas", "General Cover 2",
"MEC+SEN Cover 2", "Simultaneas Cover 2"). Por defecto NULL}
}
\description{
Esta función crea la gráfica del ratio periodo de liquidación por título en formato heatmap
}
