% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_collateral_management.R
\name{gt_gar_dep_remuneracion_diaria}
\alias{gt_gar_dep_remuneracion_diaria}
\title{Gráfica la garantía depositada vs. utilidad remuneración vs. tarifa remuneración diaria (lineas)}
\usage{
gt_gar_dep_remuneracion_diaria(datos, colores, fixedrange = FALSE)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_cm_remuneracion_resumen}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
#' por la función \code{\link{dt_adm_gen_colores}}}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}
}
\description{
Esta función crea la gráfica de la garantía depositada vs. utilidad remuneración vs. tarifa remuneración
 diaria en formato de lineas
}
