% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_collateral_management.R
\name{gt_gar_dep_remuneracion_por_miembro_fecha}
\alias{gt_gar_dep_remuneracion_por_miembro_fecha}
\title{Gráfica la garantía depositada vs. utilidad remuneración vs. tarifa remuneración por miembro de una fecha de análisis (barras)}
\usage{
gt_gar_dep_remuneracion_por_miembro_fecha(datos, colores, fixedrange = FALSE)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_cm_remuneracion_resumen}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función \code{\link{dt_adm_gen_colores}}}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}
}
\description{
Esta función crea la gráfica de la garantía depositada vs. utilidad remuneración vs. tarifa remuneración por miembro para una
fecha de análisis en formato de barras
}
