% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_collateral_management.R
\name{gt_gar_dep_vol_negociado_promedio_diario_por_accion}
\alias{gt_gar_dep_vol_negociado_promedio_diario_por_accion}
\title{Gráfica la garantía depositada vs volumen negociado por acción promedio diario (barras+puntos)}
\usage{
gt_gar_dep_vol_negociado_promedio_diario_por_accion(
  datos,
  colores,
  fixedrange = FALSE,
  boton_activo = NULL
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_cm_promedio_diario}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
#' por la función \code{\link{dt_colores}}}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico ("Efectivo", "Acciones", "Efectivo Cover 2", "Acciones Cover 2"). Por defecto NULL}
}
\description{
Esta función crea la gráfica de la garantía depositada (barras) vs volumen negociado (puntos)
por acción promedio diario
}
