% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_garantia.R
\name{gt_gar_ggl_ind_fgc_por_miembro_liq}
\alias{gt_gar_ggl_ind_fgc_por_miembro_liq}
\title{Gráfica la garantía depositada (GGL) vs. exigida (individual + FGC) por miembro liquidador (barras+puntos)}
\usage{
gt_gar_ggl_ind_fgc_por_miembro_liq(datos, colores, fixedrange = FALSE)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_gar_ggl_ind_fgc_liq}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función colores}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}
}
\description{
Esta función crea la gráfica de la garantía depositada (GGL) vs. exigida (individual + FGC) por miembro liquidador
en formato de barras y puntos.
}
