% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_ingresos.R
\name{gt_ing}
\alias{gt_ing}
\title{Gráfica los ingresos (pie)}
\usage{
gt_ing(datos, colores, boton_activo = NULL, botones_inactivos = c())
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_ing_resumen}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función colores}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico. Por defecto NULL}

\item{botones_inactivos}{clase array character. Nombre de los botones a desactivar
en la gráfica ("Segmento", "Tipo Producto", "Subtipo Producto", "Origen Producto",
"Tipo Cuenta Gar.", "Concepto Tarifa"). Por defecto c()}
}
\description{
Esta función crea la gráfica de los ingresos en formato de pie.
La información se muestra acorde a la agrupación relacionada con cada botón
}
