% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_ingresos.R
\name{gt_ing_promedio_diario_por_mes_miembro_producto}
\alias{gt_ing_promedio_diario_por_mes_miembro_producto}
\title{Gr치fica el comportamiento de los ingresos promedio diario por mes, miembro y producto (treemap)}
\usage{
gt_ing_promedio_diario_por_mes_miembro_producto(datos)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la funci칩n
\code{\link{dt_gen_ing_resumen}} o tener una estructura igual a dichos datos}
}
\description{
Esta funci칩n crea la gr치fica del comportamiento de los ingresos promedio diario por mes, miembro y producto en
formato treemap
}
