% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_ingresos.R
\name{gt_ing_resumen}
\alias{gt_ing_resumen}
\title{Tabla ingresos resumen}
\usage{
gt_ing_resumen(datos, fecha_analisis, pageLength = 100, style = "bootstrap4")
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_ing_resumen}} o tener una estructura igual a dichos datos}

\item{fecha_analisis}{clase date. Fecha en la que se realiza el análisis (Último día de los datos)}

\item{pageLength}{clase number. Número de filas por hoja que alojara
la tabla. Por defecto 100}

\item{style}{clase character. Estilo boostrap que se debe utilizar
para renderizar la tabla. Por defecto "bootstrap4"}
}
\description{
Esta función crea la tabla ingresos en formato html
}
