% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_garantia.R
\name{gt_lmc_distribucion_consumo}
\alias{gt_lmc_distribucion_consumo}
\title{Gráfica la distribución consumo LMC por miembro (boxplot)}
\usage{
gt_lmc_distribucion_consumo(datos, fixedrange = FALSE)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_lmc_consumo}} o tener una estructura igual a dichos datos}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}
}
\description{
Esta función crea la gráfica de la distribución del consumo del LMC por miembro en formato de boxplot
}
