% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_general.R
\docType{data}
\name{gt_mbb_minimizar_pantalla}
\alias{gt_mbb_minimizar_pantalla}
\title{Retorna el modeBarButton minimizar pantalla}
\format{
An object of class \code{list} of length 3.
}
\usage{
gt_mbb_minimizar_pantalla
}
\description{
Retorna el modeBarButton minimizar pantalla
}
\keyword{datasets}
