% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_general.R
\docType{data}
\name{gt_mensaje_error}
\alias{gt_mensaje_error}
\title{Retorna la gráfica mensaje error}
\format{
An object of class \code{plotly} (inherits from \code{htmlwidget}) of length 8.
}
\usage{
gt_mensaje_error
}
\description{
Retorna la gráfica mensaje error
}
\keyword{datasets}
