% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_volumen.R
\name{gt_miembros_operando}
\alias{gt_miembros_operando}
\title{Grafica el número de miembros operando en cada segmento por mes (barras)}
\usage{
gt_miembros_operando(datos, colores, fixedrange = FALSE)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_vol_resumen}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función \code{\link{dt_colores}}}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}
}
\description{
Esta función crea la gráfica el ranking de los miembros por volumen operado.
La información se muestra acorde a la agrupación relacionada con cada botón
}
