% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_garantia.R
\name{gt_pa_bruta_gar_exi_por_producto}
\alias{gt_pa_bruta_gar_exi_por_producto}
\title{Gráfica la posición abierta bruta (dos puntas) vs garantía exigida por producto (barras)}
\usage{
gt_pa_bruta_gar_exi_por_producto(
  datos,
  fixedrange = FALSE,
  boton_activo = NULL,
  botones_inactivos = c()
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_gar_exi_resumen}} o tener una estructura igual a dichos datos}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico. Por defecto NULL}

\item{botones_inactivos}{clase array character. Nombre de los botones a desactivar
en la gráfica (....). Por defecto c()}
}
\description{
Esta función crea la gráfica de la posición abierta bruta (eje 1) vs. garantía exigida (eje y2) vs.
y el ratio de cobertura (eje y3) por producto en formato de barras.
La información se muestra acorde a la agrupación relacionada con cada botón
}
