% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_garantia.R
\name{gt_pa_gar_exi_rss_por_miembro_liq}
\alias{gt_pa_gar_exi_rss_por_miembro_liq}
\title{Gráfica la posición abierta vs garantía exigida vs riesgo en situación de estrés por miembro liquidador (barras)}
\usage{
gt_pa_gar_exi_rss_por_miembro_liq(
  datos,
  fixedrange = FALSE,
  boton_activo = NULL,
  botones_inactivos = c()
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_gar_dep_exi_liq}} o tener una estructura igual a dichos datos}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico. Por defecto NULL}

\item{botones_inactivos}{clase vector character. Vector de los nombres de los botones a desactivar
en la gráfica (....). Por defecto c()}
}
\description{
Esta función crea la gráfica de la posición abierta (eje y1) vs. garantía exigida (eje y2) vs.
riesgo en situación de estrés (eje y3) por miembro liquidador en formato de barras.
La información se muestra acorde a la agrupación relacionada con cada botón
}
