% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_posicion_abierta.R
\name{gt_pa_por_miembro}
\alias{gt_pa_por_miembro}
\title{Gráfica la posicón abierta por miembro (barras)}
\usage{
gt_pa_por_miembro(
  datos,
  colores,
  fixedrange = FALSE,
  boton_activo = NULL,
  botones_inactivos = c()
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_pa_resumen}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función colores}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico. Por defecto NULL}

\item{botones_inactivos}{clase array character. Nombre de los botones a desactivar
en la gráfica ("Segmento", "Tipo Producto",
"Subtipo Producto", "Origen Producto", "Tipo Cuenta Gar."). Por defecto c()}
}
\description{
Esta función crea la gráfica de la posicón abierta (compradora eje positivo y vendedora eje negativo) por miembro
en formato de barras.
La información se muestra acorde a la agrupación relacionada con cada botón
}
