% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_posicion_abierta.R
\name{gt_pa_resumen}
\alias{gt_pa_resumen}
\title{Tabla posición abierta resumen}
\usage{
gt_pa_resumen(datos, fecha_analisis, pageLength = 100, style = "bootstrap4")
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_pa_resumen}} o tener una estructura igual a dichos datos}

\item{fecha_analisis}{clase date. Fecha en la que se realiza el análisis (Último día de los datos)}

\item{pageLength}{clase number. Número de filas por hoja que alojara
la tabla. Por defecto 100}

\item{style}{clase character. Estilo boostrap que se debe utilizar
para renderizar la tabla. Por defecto "bootstrap4"}
}
\description{
Esta función crea la tabla posición abierta en formato html
}
