% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_renta_fija.R
\name{gt_rf_curva_tes}
\alias{gt_rf_curva_tes}
\title{Grafica la curva tes (lineas + barras)}
\usage{
gt_rf_curva_tes(datos, fecha_analisis, fixedrange = FALSE)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función}

\item{fecha_analisis}{clase date. Fecha en la que se realiza el análisis (Último día de los datos)
\code{\link{dt_dv_curva_fwd}} o tener una estructura igual a dichos datos}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}
}
\description{
Esta función crea la gráfica de la curva tes en formato de lineas y barras.
La información se muestra acorde a la agrupación relacionada con cada botón
}
