% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_renta_fija.R
\name{gt_rf_dur_mod_por_miembro}
\alias{gt_rf_dur_mod_por_miembro}
\title{Grafica la duración modificada por miembro (barras)}
\usage{
gt_rf_dur_mod_por_miembro(
  datos,
  colores,
  fixedrange = FALSE,
  boton_activo = NULL
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_rf_pa_por_tramo}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función \code{\link{dt_colores}}}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico ("General", "Tramo"). Por defecto NULL}
}
\description{
Esta función crea la gráfica de la duración modificada por miembro
en formato de barras.
La información se muestra acorde a la agrupación relacionada con cada botón
}
