% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_renta_fija.R
\name{gt_rf_pa_neta_diaria_por_tramo}
\alias{gt_rf_pa_neta_diaria_por_tramo}
\title{Grafica la posición abierta neta diaria por tramo  (lineas)}
\usage{
gt_rf_pa_neta_diaria_por_tramo(datos, colores, fixedrange = FALSE)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_rf_pa_por_tramo}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función \code{\link{dt_colores}}}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}
}
\description{
Esta función crea la gráfica de la posición abierta neta diaria por tramo
en formato de lineas
}
