% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_rss.R
\name{gt_rss_fgc_diario}
\alias{gt_rss_fgc_diario}
\title{Grafica el fgc diario (lineas)}
\usage{
gt_rss_fgc_diario(datos, colores, fixedrange = FALSE)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_rss_fgc}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
#' por la función colores}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}
}
\description{
Esta función crea la gráfica del fgc diario en formato de lineas
}
