% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_rss.R
\name{gt_rss_fgc_resumen}
\alias{gt_rss_fgc_resumen}
\title{Tabla fgc resumen}
\usage{
gt_rss_fgc_resumen(datos, pageLength = 100, style = "bootstrap4")
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_rss_fgc}} o tener una estructura igual a dichos datos}

\item{pageLength}{clase number. Número de filas por hoja que alojara
la tabla. Por defecto 100}

\item{style}{clase character. Estilo boostrap que se debe utilizar
para renderizar la tabla. Por defecto "bootstrap4"}
}
\description{
Esta función crea la tabla fgc en formato html
}
