% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_rss.R
\name{gt_rss_promedio}
\alias{gt_rss_promedio}
\title{Gráfica el riesgo en situación de estres promedio 90 dias vs FGC (lineas)}
\usage{
gt_rss_promedio(
  datos,
  fixedrange = FALSE,
  boton_activo = NULL,
  botones_inactivos = c()
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_rss_promedio}} o tener una estructura igual a dichos datos}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico. Por defecto NULL}

\item{botones_inactivos}{clase array character. Nombre de los botones a desactivar
en la gráfica (....). Por defecto c()}
}
\description{
Esta función crea la gráfica del riesgo en situación de estres promedio 90 dias vs FGC en formato lineas.
La información se muestra acorde a la agrupación relacionada con cada botón
}
