% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_renta_variable.R
\name{gt_rv_activos_elegibles_repo}
\alias{gt_rv_activos_elegibles_repo}
\title{Tabla de los activos elegibles repo}
\usage{
gt_rv_activos_elegibles_repo(
  datos,
  fecha_analisis,
  pageLength = 100,
  style = "bootstrap4"
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_ing_cumplimiento_presupuesto}} o tener una estructura igual a dichos datos}

\item{fecha_analisis}{clase date. Fecha en la que se realiza el análisis (Último día de los datos)}

\item{pageLength}{clase number. Número de filas por hoja que alojara
la tabla. Por defecto 100}

\item{style}{clase character. Estilo boostrap que se debe utilizar
para renderizar la tabla. Por defecto "bootstrap4"}
}
\description{
Esta función crea la tabla de los activos elegibles repo en formato html.
}
