% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_renta_variable.R
\name{gt_rv_pa_repos_por_activo_miembro_titular_garantia}
\alias{gt_rv_pa_repos_por_activo_miembro_titular_garantia}
\title{Gráfica la posición abierta por activo, miembro y titular garantía (treemap)}
\usage{
gt_rv_pa_repos_por_activo_miembro_titular_garantia(datos)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_rv_pa_repos}} o tener una estructura igual a dichos datos}
}
\description{
Esta función crea la gráfica de la  posición abierta por activo, miembro y titular garantía en
formato treemap
}
