% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_renta_variable.R
\name{gt_rv_pa_volumen_por_activo_diario}
\alias{gt_rv_pa_volumen_por_activo_diario}
\title{Grafica el volumen operado por activo (lineas)}
\usage{
gt_rv_pa_volumen_por_activo_diario(datos, fixedrange = FALSE)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_rv_vol}} o tener una estructura igual a dichos datos}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}
}
\description{
Esta función crea la gráfica el volumen operado por activo (eje y1), la disperción VMD (eje y2) y el porcentaje de marcación (eje y3)
en formato de lineas
}
