% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_renta_variable.R
\name{gt_rv_volumen_diario}
\alias{gt_rv_volumen_diario}
\title{Grafica el volumen operado (lineas)}
\usage{
gt_rv_volumen_diario(datos, fixedrange = FALSE)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_rv_vol}} o tener una estructura igual a dichos datos}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}
}
\description{
Esta función crea la gráfica el volumen operado (eje y1) y el volumen operado promedio movil (eje y2)
en formato de lineas
}
