% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_tarifas.R
\name{gt_tar_cumplimiento_presupuesto_resumen}
\alias{gt_tar_cumplimiento_presupuesto_resumen}
\title{Tabla cumplimiento presupuesto resumen}
\usage{
gt_tar_cumplimiento_presupuesto_resumen(
  datos,
  fecha_analisis,
  pageLength = 100,
  style = "bootstrap4"
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_tar_cumplimiento_presupuesto}} o tener una estructura igual a dichos datos}

\item{fecha_analisis}{clase date. Fecha en la que se realiza el análisis (Último día de los datos)}

\item{pageLength}{clase number. Número de filas por hoja que alojara
la tabla. Por defecto 100}

\item{style}{clase character. Estilo boostrap que se debe utilizar
para renderizar la tabla. Por defecto "bootstrap4"}
}
\description{
Esta función crea la tabla cumlimplimiento presupuesto en formato html.
}
