% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_tarifas.R
\name{gt_tar_por_miembro}
\alias{gt_tar_por_miembro}
\title{Grafica las tarifas por miembro  (barras)}
\usage{
gt_tar_por_miembro(
  datos,
  colores,
  fixedrange = FALSE,
  boton_activo = NULL,
  botones_inactivos = c()
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_tar_resumen}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función \code{\link{dt_adm_gen_colores}}}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico. Por defecto NULL}

\item{botones_inactivos}{clase vector character. Vector de los nombres de los botones a desactivar
en la gráfica ("Segmento", "Tipo Producto",
"Subtipo Producto", "Origen Producto", "Tipo Cuenta Gar.", "Concepto Tarifa"). Por defecto c()}
}
\description{
Esta función crea la gráfica de las tarifas por miembro en formato de barras.
La información se muestra acorde a la agrupación relacionada con cada botón
}
