% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_tarifas.R
\name{gt_tar_promedio_diario_por_mes_producto_miembro}
\alias{gt_tar_promedio_diario_por_mes_producto_miembro}
\title{Gr치fica el comportamiento de las tarifas promedio diario por mes, producto y miembro (treemap)}
\usage{
gt_tar_promedio_diario_por_mes_producto_miembro(datos)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la funci칩n
\code{\link{dt_gen_tar_resumen}} o tener una estructura igual a dichos datos}
}
\description{
Esta funci칩n crea la gr치fica del comportamiento de las tarifas promedio diario por mes, producto y miembro en
formato treemap
}
