% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_volumen.R
\name{gt_vol}
\alias{gt_vol}
\title{Grafica el volumen operado (barras h)}
\usage{
gt_vol(
  datos,
  colores,
  fixedrange = FALSE,
  boton_activo = NULL,
  botones_inactivos = c()
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_vol_resumen}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función \code{\link{dt_colores}}}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico. Por defecto NULL}

\item{botones_inactivos}{clase vector character. Vector de los nombres de los botones a desactivar
en la gráfica ("Segmento", "Tipo Producto", "Subtipo Producto",
"Origen Producto", "Tipo Miembro", "Tipo Cuenta Gar."). Por defecto c()}
}
\description{
Esta función crea la gráfica del volumen operado (compra y venta)
en formato de barras horizantales.
La información se muestra acorde a la agrupación relacionada con cada botón
}
