% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_volumen.R
\name{gt_vol_diario}
\alias{gt_vol_diario}
\title{Gráfica el volumen operado diario (lineas)}
\usage{
gt_vol_diario(
  datos,
  colores,
  fixedrange = FALSE,
  dash_board = TRUE,
  boton_activo = NULL,
  botones_inactivos = c()
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_vol_resumen}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
#' por la función colores}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{dash_board}{clase boolean. TRUE si se desea monstrar en la agrupaciones simetricas unicamente el eje positivo. Por defecto TRUE}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico. Por defecto NULL}

\item{botones_inactivos}{clase array character. Nombre de los botones a desactivar
en la gráfica ("Segmento", "Tipo Producto", "Subtipo Producto",
"Origen Producto", "Tipo Miembro", "Tipo Cuenta Gar."). Por defecto c()}
}
\description{
Esta función crea la gráfica del volumen operado diario (compra eje positivo y venta eje negativo) en formato de lineas.
La información se muestra acorde a la agrupación relacionada con cada botón
}
