% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_volumen.R
\name{gt_vol_promedio_diario_tipocuenta}
\alias{gt_vol_promedio_diario_tipocuenta}
\title{Grafica el volumen operado promedio diario por (Mes o Año) y tipo de cuenta (barras)}
\usage{
gt_vol_promedio_diario_tipocuenta(
  datos,
  colores,
  fixedrange = FALSE,
  promedio = "m",
  boton_activo = NULL
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_vol_resumen}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función \code{\link{dt_adm_gen_colores}}}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{promedio}{clase character. "m" si se desea promediar por mes y "y" si se desea promediar por año. Por defecto "m"}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico ("Posición Propia", "Posición Terceros", "Posición Cartera"). Por defecto NULL}
}
\description{
Esta función crea la gráfica del volumen operado promedio diario (compra eje positivo y venta eje negativo)
por tipo de cuenta en formato de barras. La información se muestra acorde a la agrupación relacionada con cada botón
}
