% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_volumen.R
\name{gt_vol_ranking_miembros}
\alias{gt_vol_ranking_miembros}
\title{Grafica el ranking de los miembros por volumen operado (barras)}
\usage{
gt_vol_ranking_miembros(
  datos,
  colores,
  fixedrange = FALSE,
  boton_activo = NULL
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_vol_resumen}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función \code{\link{dt_adm_gen_colores}}}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico ("General", "Derivados OTC", "Divisas",
"Estandarizado", "C. Renta Variable", "Repos y TTV", "Simultaneas"). Por defecto NULL}
}
\description{
Esta función crea la gráfica el ranking de los miembros por volumen operado en formato de barras.
La información se muestra acorde a la agrupación relacionada con cada botón
}
