% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_general.R
\name{dt_abrir_conexion}
\alias{dt_abrir_conexion}
\title{Abrir conexion de la base de datos}
\usage{
dt_abrir_conexion(config)
}
\arguments{
\item{config}{clase data.frame. Configuración de la conexión}
}
\description{
Esta función abre la conexion de la base de datos
}
