% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_volumen.R
\name{dt_gen_vol_resumen}
\alias{dt_gen_vol_resumen}
\title{Descarga los datos gen_volumen_resumen}
\usage{
dt_gen_vol_resumen(
  conexion,
  periodo_analisis = NULL,
  fecha_analisis = NULL,
  ficticio = FALSE,
  segmentos = NULL,
  oracle = FALSE
)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{periodo_analisis}{clase array date. Debe contener la fecha inicio y fin del análisis}

\item{fecha_analisis}{clase date. Debe contener la fecha del análisis, si el parametro periodo_analisis es
diferente de NULL este parametro no se tendra en cuenta. Por defecto NULL}

\item{ficticio}{clase boolean. TRUE si se desea que el "ID_SEUDONIMO" de los miembros se igua al "ID_FICTICIO"  en
caso contrario sera igual al "ID". Por defecto FALSE}

\item{segmentos}{clase array character. Lista de segmentos de los cuales se desea descargar la información.
Por defecto descarga la información de todos los segmentos.}

\item{oracle}{clase boolean. TRUE si la base de datos utilizada es oracle. Por defecto FALSE.}
}
\description{
Esta función descarga los datos de la tabla gen_volumen_resumen para un periodo de análisis y
con base en los parametros ingresados
}
