% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_renta_variable.R
\name{dt_rv_pa_repos}
\alias{dt_rv_pa_repos}
\title{Descarga los datos gen_posicion_abierta para el producto repos}
\usage{
dt_rv_pa_repos(
  conexion,
  proveedor = "MySQL",
  periodo_analisis = NULL,
  fecha_analisis = NULL,
  ficticio = FALSE,
  nominal = FALSE,
  valorada = TRUE,
  importe = FALSE
)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{proveedor}{clase character. Proveedor de la base de datos ("Oracle", "MySQL"). Por defecto "MySQL"}

\item{periodo_analisis}{clase array date. Debe contener la fecha inicio y fin del análisis}

\item{fecha_analisis}{clase date. Debe contener la fecha del análisis, si el parametro periodo_analisis es
diferente de NULL este parametro no se tendra en cuenta. Por defecto NULL}

\item{ficticio}{clase boolean. TRUE si se desea que el "ID_SEUDONIMO" de los miembros se igua al "ID_FICTICIO"  en
caso contrario sera igual al "ID". Por defecto FALSE}

\item{nominal}{clase boolean. TRUE si se desea descargar la posición abierta nominal Por defecto TRUE}

\item{valorada}{clase boolean. TRUE si se desea descargar la posición abierta valorada. Por defecto TRUE}

\item{importe}{clase boolean. TRUE si se desea descargar la posición abierta importe. Por defecto FALSE}
}
\description{
Esta función descarga los datos de la tabla gen_posicion_abierta para el producto repos, un periodo de análisis y
con base en los parametros ingresados
}
