% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_collateral_management.R
\name{gt_gar_dep}
\alias{gt_gar_dep}
\title{Gráfica la garantía depositada (pie)}
\usage{
gt_gar_dep(datos, colores)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_gar_dep_resumen_fecha}} o tener una estructura igual a dichos datos}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función colores}
}
\description{
Esta función crea la gráfica de la garantía depositada tipo activo en formato de pie
}
