% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_ingresos.R
\name{gt_ing_cumplimiento_presupuesto}
\alias{gt_ing_cumplimiento_presupuesto}
\title{Gráfica el resumen del cumplimiento del presupuesto}
\usage{
gt_ing_cumplimiento_presupuesto(
  datos,
  fixedrange = FALSE,
  boton_activo = NULL,
  botones_inactivos = c()
)
}
\arguments{
\item{datos}{clase data.frame. Los datos deben ser los generados por la función
\code{\link{dt_gen_ing_cumplimiento_presupuesto}} o tener una estructura igual a dichos datos}

\item{fixedrange}{clase boolean. TRUE si se desea desactivar la función de zoom en las gráficas. Por defecto FALSE}

\item{boton_activo}{clase character. Si se desea que la gráfica se inicialice
con un botón seleccionado en especifico. Por defecto NULL}

\item{botones_inactivos}{clase array character. Nombre de los botones a desactivar
en la gráfica ("Segmento", "Tipo Producto", "Origen Producto", "Producto"). Por defecto c()}

\item{colores}{clase data.frame. Debe contener los datos generados
por la función colores}
}
\description{
Esta función crea la gráfica resumen del cumplimiento del presupuesto para el utlimo mes
y periodo(fecha min a fecha max de los datos) (barras).
La información se muestra acorde la agrupación relacionada con cada botón
}
